Fix data file module inference
state.json is not a data_file. Not excluding it can lead to an error when trying to load a dataset with a split that has only one shard.

Excluding state.json should be fine because it is excluded later on anyway when the format is inferred as e.g. "arrow"